pyproject.toml
[tool.poetry]
name = "youtube-subscriptions"
version = "1.0.0"
description = "YouTube Subscriptions Monitor API"
authors = ["azcoigreach"]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "*"
httpx = "*"
google-auth = "*"
google-auth-oauthlib = "*"
google-api-python-client = "*"
python-dotenv = "*"
uvicorn = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
